+++ momentum-dao-be/src/main/java/com/dao/momentum/organization/employee/query/controller/AdminEmployeeQueryController.java
@@ -6,7 +6,7 @@ import com.dao.momentum.organization.employee.query.dto.request.EmployeeSearchRe
 import com.dao.momentum.organization.employee.query.dto.response.AppointListResponse;
 import com.dao.momentum.organization.employee.query.dto.response.EmployeeDetailsResponse;
 import com.dao.momentum.organization.employee.query.dto.response.EmployeeListResponse;
-import com.dao.momentum.organization.employee.query.service.AdminEmployeeQueryService;
+import com.dao.momentum.organization.employee.query.service.EmployeeQueryService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
@@ -18,7 +18,7 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 @Tag(name = "사원 정보 조회", description = "관리자 사원 정보 조회 API")
 public class AdminEmployeeQueryController {
-    private final AdminEmployeeQueryService adminEmployeeQueryService;
+    private final EmployeeQueryService employeeQueryService;
 
     @GetMapping
     @Operation(summary = "사원 전체 목록 조회", description = "관리자가 회사의 사원 전체 목록을 조회합니다.")
@@ -28,7 +28,7 @@ public class AdminEmployeeQueryController {
 
         return ResponseEntity.ok(
                 ApiResponse.success(
-                        adminEmployeeQueryService.getEmployees(employeeSearchRequest)
+                        employeeQueryService.getEmployees(employeeSearchRequest)
                 )
         );
     }
@@ -41,7 +41,7 @@ public class AdminEmployeeQueryController {
 
         return ResponseEntity.ok(
                 ApiResponse.success(
-                        adminEmployeeQueryService.getEmployeeDetails(empId)
+                        employeeQueryService.getEmployeeDetails(empId)
                 )
         );
     }
@@ -54,7 +54,7 @@ public class AdminEmployeeQueryController {
 
         return ResponseEntity.ok(
                 ApiResponse.success(
-                        adminEmployeeQueryService.getAppoints(appointSearchRequest)
+                        employeeQueryService.getAppoints(appointSearchRequest)
                 )
         );
     }
