+++ momentum-dao-be/src/main/java/com/dao/momentum/organization/employee/query/controller/EmployeeQueryController.java
@@ -2,7 +2,7 @@ package com.dao.momentum.organization.employee.query.controller;
 
 import com.dao.momentum.common.dto.ApiResponse;
 import com.dao.momentum.organization.employee.query.dto.response.EmployeeDetailsResponse;
-import com.dao.momentum.organization.employee.query.service.AdminEmployeeQueryService;
+import com.dao.momentum.organization.employee.query.service.EmployeeQueryService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
@@ -18,18 +18,18 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 @Tag(name = "사원 정보 조회", description = "사원 본인 정보 조회 API")
 public class EmployeeQueryController {
-    private final AdminEmployeeQueryService adminEmployeeQueryService;
+    private final EmployeeQueryService employeeQueryService;
 
     @GetMapping("/me")
     @Operation(summary = "사원 개인 정보 조회", description = "사원이 본인 정보를 조회합니다.")
-    public ResponseEntity<ApiResponse<EmployeeDetailsResponse>> getEmployeeDetails(
+    public ResponseEntity<ApiResponse<EmployeeDetailsResponse>> getMyInfo(
             @AuthenticationPrincipal UserDetails userDetails
             ) {
         long empId = Long.parseLong(userDetails.getUsername());
 
         return ResponseEntity.ok(
                 ApiResponse.success(
-                        adminEmployeeQueryService.getEmployeeDetails(empId)
+                        employeeQueryService.getEmployeeDetails(empId)
                 )
         );
     }
